# docs/openapi/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "job-tracker-api",
    "description": "REST API for job-tracker-dashboard",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/loops": {
      "get": {
        "tags": [
          "loops"
        ],
        "summary": "List loops",
        "operationId": "list_loops_api_v1_loops_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_archived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Archived"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "loops"
        ],
        "summary": "Create loop",
        "operationId": "create_loop_api_v1_loops_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoopCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}": {
      "get": {
        "tags": [
          "loops"
        ],
        "summary": "Get loop",
        "operationId": "get_loop_api_v1_loops__loop_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Loop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "loops"
        ],
        "summary": "Update loop",
        "operationId": "patch_loop_api_v1_loops__loop_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Loop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoopUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "loops"
        ],
        "summary": "Archive loop",
        "operationId": "delete_loop_api_v1_loops__loop_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Loop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoopRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/loops/{loop_id}/matches/{match_id}/analyses": {
      "post": {
        "tags": [
          "vacancy-analysis"
        ],
        "summary": "Create and save vacancy analysis",
        "operationId": "create_analysis_api_v1_loops__loop_id__matches__match_id__analyses_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancyAnalysisCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "vacancy-analysis"
        ],
        "summary": "List saved vacancy analyses for a match",
        "operationId": "list_analyses_api_v1_loops__loop_id__matches__match_id__analyses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyAnalysisListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loops/{loop_id}/matches/{match_id}/analyses/latest": {
      "get": {
        "tags": [
          "vacancy-analysis"
        ],
        "summary": "Get latest saved vacancy analysis for a match",
        "operationId": "latest_analysis_api_v1_loops__loop_id__matches__match_id__analyses_latest_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "loop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Loop Id"
            }
          },
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacancyAnalysisRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoopCreate": {
        "properties": {
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "excluded_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Keywords"
          },
          "employment_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Employment Types"
          },
          "work_modes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Work Modes"
          },
          "selected_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Sources"
          },
          "auto_discovery_enabled": {
            "type": "boolean",
            "title": "Auto Discovery Enabled",
            "default": false
          },
          "discovery_radius_km": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 250.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovery Radius Km"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "target_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "radius_km": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Radius Km"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "archived"
            ],
            "title": "Status",
            "default": "active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "LoopCreate"
      },
      "LoopListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LoopRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "LoopListResponse"
      },
      "LoopRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "target_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "radius_km": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radius Km"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          },
          "excluded_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Excluded Keywords"
          },
          "employment_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Employment Types"
          },
          "work_modes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Work Modes"
          },
          "selected_sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Sources"
          },
          "auto_discovery_enabled": {
            "type": "boolean",
            "title": "Auto Discovery Enabled"
          },
          "discovery_radius_km": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovery Radius Km"
          },
          "last_discovery_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Discovery At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "target_role",
          "location",
          "radius_km",
          "sources",
          "status",
          "keywords",
          "excluded_keywords",
          "employment_types",
          "work_modes",
          "selected_sources",
          "auto_discovery_enabled",
          "discovery_radius_km",
          "last_discovery_at",
          "created_at",
          "updated_at"
        ],
        "title": "LoopRead"
      },
      "LoopUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "target_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Role"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "radius_km": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Radius Km"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "paused",
                  "archived"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "excluded_keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Keywords"
          },
          "employment_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employment Types"
          },
          "work_modes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Modes"
          },
          "selected_sources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Sources"
          },
          "auto_discovery_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Discovery Enabled"
          },
          "discovery_radius_km": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 250.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovery Radius Km"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LoopUpdate"
      },
...
      "VacancyAnalysisCreate": {
        "properties": {
          "analysis_type": {
            "type": "string",
            "enum": [
              "basic",
              "ai"
            ],
            "title": "Analysis Type",
            "default": "basic"
          },
          "resume_text": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Resume Text"
          },
          "language": {
            "type": "string",
            "maxLength": 8,
            "minLength": 2,
            "title": "Language",
            "default": "ru"
          },
          "include_cover_letter": {
            "type": "boolean",
            "title": "Include Cover Letter",
            "default": false
          },
          "include_interview_questions": {
            "type": "boolean",
            "title": "Include Interview Questions",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "resume_text"
        ],
        "title": "VacancyAnalysisCreate"
      },
      "VacancyAnalysisListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VacancyAnalysisRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "VacancyAnalysisListResponse"
      },
...
      "VacancyAnalysisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "loop_id": {
            "type": "string",
            "title": "Loop Id"
          },
          "match_id": {
            "type": "string",
            "format": "uuid",
            "title": "Match Id"
          },
          "analysis_type": {
            "type": "string",
            "enum": [
              "basic",
              "ai"
            ],
            "title": "Analysis Type"
          },
          "provider": {
            "type": "string",
            "enum": [
              "deterministic",
              "ollama",
              "openai",
              "gemini",
              "groq"
            ],
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "basic",
              "premium"
            ],
            "title": "Plan"
          },
          "resume_hash": {
            "type": "string",
            "title": "Resume Hash"
          },
          "vacancy_snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Vacancy Snapshot"
          },
          "overall_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Overall Score"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Gaps"
          },
          "risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risks"
          },
          "recommended_cv_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Cv Keywords"
          },
          "application_angle": {
            "type": "string",
            "title": "Application Angle"
          },
          "cover_letter_draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Letter Draft"
          },
          "interview_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Interview Questions"
          },
          "model_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Model Info"
          },
          "quota_day": {
            "type": "string",
            "format": "date",
            "title": "Quota Day"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "quota": {
            "$ref": "#/components/schemas/VacancyAnalysisQuota"
          }
        },
        "type": "object",
        "required": [
          "id",
          "loop_id",
          "match_id",
          "analysis_type",
          "provider",
          "model",
          "plan",
          "resume_hash",
          "vacancy_snapshot",
          "overall_score",
          "summary",
          "strengths",
          "gaps",
          "risks",
          "recommended_cv_keywords",
          "application_angle",
          "cover_letter_draft",
          "interview_questions",
          "model_info",
          "quota_day",
          "created_at",
          "quota"
        ],
        "title": "VacancyAnalysisResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}